Create new folders for api and info page markdown functions
