+++ packages/docusaurus-plugin-openapi-docs/src/markdown/api-page/createParamsDetails.ts
@@ -5,10 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  * ========================================================================== */
 
-import { ApiItem } from "../types";
+import { ApiItem } from "../../types";
+import { create } from "../utils";
 import { createDetails } from "./createDetails";
 import { createDetailsSummary } from "./createDetailsSummary";
-import { create } from "./utils";
 
 interface Props {
   parameters: ApiItem["parameters"];
