+++ packages/docusaurus-plugin-openapi-docs/src/markdown/api-page/createParamsTable.ts
@@ -7,11 +7,11 @@
 
 import { escape } from "lodash";
 
-import { ApiItem } from "../types";
+import { ApiItem } from "../../types";
+import { getQualifierMessage, getSchemaName } from "../schema";
+import { create, guard } from "../utils";
 import { createDescription } from "./createDescription";
 import { createFullWidthTable } from "./createFullWidthTable";
-import { getQualifierMessage, getSchemaName } from "./schema";
-import { create, guard } from "./utils";
 
 interface Props {
   parameters: ApiItem["parameters"];
