+++ packages/docusaurus-plugin-openapi-docs/src/markdown/api-page/createRequestBodyDetails.ts
