+++ packages/docusaurus-plugin-openapi-docs/src/markdown/api-page/createStatusCodes.ts
@@ -5,10 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  * ========================================================================== */
 
-import { ApiItem } from "../types";
+import { ApiItem } from "../../types";
+import { create } from "../utils";
 import { createDescription } from "./createDescription";
 import { createSchemaDetails } from "./createSchemaDetails";
-import { create } from "./utils";
 
 interface Props {
   responses: ApiItem["responses"];
