+++ packages/docusaurus-plugin-openapi-docs/src/markdown/info-page/createLicense.ts
@@ -1,5 +1,12 @@
-import { create } from "../utils";
+/* ============================================================================
+ * Copyright (c) Palo Alto Networks
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ * ========================================================================== */
+
 import { LicenseObject } from "../../openapi/types";
+import { create } from "../utils";
 
 export function createLicense(license: LicenseObject) {
   if (!license) return "";
