+++ packages/docusaurus-plugin-openapi-docs/src/markdown/info-page/createTermsOfService.ts
@@ -1,3 +1,10 @@
+/* ============================================================================
+ * Copyright (c) Palo Alto Networks
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ * ========================================================================== */
+
 import { create } from "../utils";
 
 export function createTermsOfService(termsOfService: string | undefined) {
